refactor: Http 요청, 응답을 기반으로 동작하는 클래스를 catalina로 이동
