+++ tomcat/src/main/java/com/techcourse/Application.java
@@ -3,8 +3,8 @@ package com.techcourse;
 import com.techcourse.controller.HomeController;
 import com.techcourse.controller.LoginController;
 import com.techcourse.controller.RegisterController;
-import org.apache.catalina.startup.Tomcat;
-import org.apache.coyote.http11.RequestMapping;
+import org.apache.catalina.RequestMapping;
+import org.apache.coyote.http11.startup.Tomcat;
 
 public class Application {
