+++ tomcat/src/main/java/org/apache/catalina/StaticResource.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.catalina;
 
 import java.io.File;
 import java.io.IOException;
