+++ tomcat/src/main/java/org/apache/coyote/http11/startup/Tomcat.java
@@ -1,10 +1,9 @@
-package org.apache.catalina.startup;
+package org.apache.coyote.http11.startup;
 
+import java.io.IOException;
 import org.apache.catalina.connector.Connector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
 
 public class Tomcat {
